# igraci_dinama_openAPI.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API igrača nogometnog kluba Dinamo",
      "description": "This API allows clients to interact with and manage information about football players including their previous clubs and stats.",
      "version": "1.0.0",
      "contact": {
        "name": "Tech Support",
        "email": "support@dinamofc.com",
        "url": "http://igraci-dinamo.com/support"
      },
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "paths": {
      "/players": {
        "get": {
          "summary": "Retrieve all players",
          "description": "Returns a list of all players with their detailed information and club associations.",
          "responses": {
            "200": {
              "description": "A list of players",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Player"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "post": {
          "summary": "Add a new player",
          "description": "Creates a new player with the given details and club information.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Player created"
            },
            "500": {
              "description": "Error creating a player"
            }
          }
        }
      },
      "/players/{broj}": {
        "get": {
          "summary": "Get a player by jersey number",
          "description": "Returns detailed information about a player identified by their jersey number.",
          "parameters": [
            {
              "name": "broj",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              },
              "description": "Jersey number of the desired player"
            }
          ],
          "responses": {
            "200": {
              "description": "Player details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            },
            "404": {
              "description": "Player not found"
            },
            "500": {
              "description": "Error fetching player"
            }
          }
        },
        "put": {
          "summary": "Update player information",
          "description": "Updates the information for the player specified by their jersey number.",
          "parameters": [
            {
              "name": "broj",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Player updated"
            },
            "404": {
              "description": "Player not found"
            },
            "500": {
              "description": "Error updating player"
            }
          }
        },
        "delete": {
          "summary": "Delete a player",
          "description": "Deletes the player with the specified jersey number from the database.",
          "parameters": [
            {
              "name": "broj",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Player deleted"
            },
            "404": {
              "description": "Player not found"
            },
            "500": {
              "description": "Error deleting player"
            }
          }
        }
      },
      "/players/position/{position}": {
        "get": {
          "summary": "Retrieve players by position",
          "description": "Returns a list of players who play in the specified position.",
          "parameters": [
            {
              "name": "position",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "Football position of the players to retrieve"
            }
          ],
          "responses": {
            "200": {
              "description": "List of players",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Player"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "No players found"
            },
            "500": {
              "description": "Error retrieving players"
            }
          }
        }
      },
      "/players/appearances/{minAppearances}": {
        "get": {
          "summary": "Retrieve players by minimum appearances",
          "description": "Returns a list of players who have appeared in at least the specified number of matches.",
          "parameters": [
            {
              "name": "minAppearances",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              },
              "description": "Minimum number of appearances to filter players"
            }
          ],
          "responses": {
            "200": {
              "description": "List of players",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Player"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "No players found"
            },
            "500": {
              "description": "Error retrieving players"
            }
          }
        }
      },
      "/players/goals/{minGoals}": {
        "get": {
            "summary": "Retrieve players by minimum goals",
            "description": "Returns a list of players who have scored at least the specified number of goals.",
            "parameters": [
            {
                "name": "minGoals",
                "in": "path",
                "required": true,
                "schema": {
                "type": "integer"
                },
                "description": "Minimum number of goals to filter players"
            }
            ],
            "responses": {
            "200": {
                "description": "List of players",
                "content": {
                "application/json": {
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/components/schemas/Player"
                    }
                    }
                }
                }
            },
            "404": {
                "description": "No players found"
            },
            "500": {
                "description": "Error retrieving players"
            }
            }
        }
        }
    },
    "components": {
      "schemas": {
        "Player": {
          "type": "object",
          "properties": {
            "ime": {
              "type": "string"
            },
            "prezime": {
              "type": "string"
            },
            "broj": {
              "type": "integer"
            },
            "pozicija": {
              "type": "string"
            },
            "datum_rodenja": {
              "type": "string",
              "format": "date"
            },
            "mjesto_rodenja": {
              "type": "string"
            },
            "visina": {
              "type": "integer"
            },
            "broj_nastupa": {
              "type": "integer"
            },
            "broj_golova": {
              "type": "integer"
            },
            "klubovi": {
              "type": "string"
            }
          }
        }
      }
    }
  }
  